requirements.txt
Django==4.1.11
djangorestframework>=3.13.1,<3.14
python-dotenv==1.0.0
drf-spectacular>=0.15.1,<0.16